requirements.txt
python-jobspy>=1.1.80
gspread>=6.0.0
pandas>=2.0.0
google-auth>=2.0.0
